fix: error suppression
Errors are throw from the plugin when a tab is trying to be accessed that doesn't exist. This could possibly be caused by users quicky closing tabs after opening them but the errors show up in the extentions menu which doesn't look nice. This commit converts the errors to normal console messages so they don't appear.